Refactor devices build
Remove #ifdefs from device drivers and compile them only when
selected. Once there, move device detection code in configure.ac
to live in one place.
