+++ common/device.c
@@ -20,12 +20,12 @@
 #include "gnokii-internal.h"
 #include "device.h"
 #include "devices/irda.h"
-#include "devices/unixbluetooth.h"
-#include "devices/tcp.h"
-#include "devices/serial.h"
-#include "devices/tekram.h"
+#include "devices/bluetooth.h"
 #include "devices/dku2libusb.h"
+#include "devices/serial.h"
 #include "devices/socketphonet.h"
+#include "devices/tcp.h"
+#include "devices/tekram.h"
 
 #include <errno.h>
 #include <sys/wait.h>
