+++ include/devices/socketphonet.h
@@ -18,12 +18,44 @@
 #ifndef _gnokii_devices_linuxphonet_h
 #define _gnokii_devices_linuxphonet_h
 
+#include "config.h"
 #include "gnokii.h"
+
+#ifdef HAVE_SOCKETPHONET
 
 int socketphonet_close(struct gn_statemachine *state);
 int socketphonet_open(const char *iface, int with_async, struct gn_statemachine *state);
 size_t socketphonet_read(int fd, __ptr_t buf, size_t nbytes, struct gn_statemachine *state);
 size_t socketphonet_write(int fd, const __ptr_t buf, size_t n, struct gn_statemachine *state);
 int socketphonet_select(int fd, struct timeval *timeout, struct gn_statemachine *state);
+
+#else
+
+int socketphonet_close(struct gn_statemachine *state)
+{
+	return -1;
+}
+
+int socketphonet_open(const char *iface, int with_async, struct gn_statemachine *state)
+{
+	return -1;
+}
+
+size_t socketphonet_read(int fd, __ptr_t buf, size_t nbytes, struct gn_statemachine *state)
+{
+	return -1;
+}
+
+size_t socketphonet_write(int fd, const __ptr_t buf, size_t n, struct gn_statemachine *state)
+{
+	return -1;
+}
+
+int socketphonet_select(int fd, struct timeval *timeout, struct gn_statemachine *state)
+{
+	return -1;
+}
+
+#endif
 
 #endif /* _gnokii_devices_linuxphonet_h */
